requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.1
pandas==0.25.3
pyparsing==2.4.6
python-dateutil==2.8.1
python-magic-bin==0.4.14
pytz==2019.3
rope==0.14.0
scipy==1.4.1
seaborn==0.9.0
six==1.13.0
xlrd==1.2.0
